requirements.txt
# Base Requirements for SiteScanner

torch>=1.10.0
# torch-scatter
# torch-sparse
# torch-cluster
# torch-spline-conv
torch_geometric>=2.0.0
numpy>=1.19.0
biopython>=1.78
e3nn>=0.4.0
scikit-learn>=1.0.0
scipy
joblib
transformers
fair-esm

# Optional: Add development/testing specific libraries below if needed
pytest>=6.0.0
pytest-mock>=3.0.0